# api/openapi/apiserver/v1/rca-complete.swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RCA API Complete",
    "version": "v1",
    "description": "Complete API surface scanned from Gin route registrations with RBAC hints."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/v1/permissions": {
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "ListPermissions",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `ListPermissions`.",
        "operationId": "ListPermissions_GET_v1_permissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "ListPermissions",
        "x-source-file": "internal/apiserver/handler/rbac.go"
      },
      "post": {
        "tags": [
          "RBAC"
        ],
        "summary": "UpsertPermission",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `UpsertPermission`.",
        "operationId": "UpsertPermission_POST_v1_permissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "UpsertPermission",
        "x-source-file": "internal/apiserver/handler/rbac.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_id": {
                    "type": "string"
                  },
                  "resource": {
                    "type": "string"
                  },
                  "action": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/permissions/{id}": {
      "delete": {
        "tags": [
          "RBAC"
        ],
        "summary": "DeletePermission",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `DeletePermission`.",
        "operationId": "DeletePermission_DELETE_v1_permissions_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "DeletePermission",
        "x-source-file": "internal/apiserver/handler/rbac.go"
      },
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "GetPermission",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `GetPermission`.",
        "operationId": "GetPermission_GET_v1_permissions_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "GetPermission",
        "x-source-file": "internal/apiserver/handler/rbac.go"
      },
      "put": {
        "tags": [
          "RBAC"
        ],
        "summary": "UpsertPermission",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `UpsertPermission`.",
        "operationId": "UpsertPermission_PUT_v1_permissions_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "UpsertPermission",
        "x-source-file": "internal/apiserver/handler/rbac.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_id": {
                    "type": "string"
                  },
                  "resource": {
                    "type": "string"
                  },
                  "action": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/roles": {
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "ListRoles",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `ListRoles`.",
        "operationId": "ListRoles_GET_v1_roles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "ListRoles",
        "x-source-file": "internal/apiserver/handler/rbac.go"
      },
      "post": {
        "tags": [
          "RBAC"
        ],
        "summary": "UpsertRole",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `UpsertRole`.",
        "operationId": "UpsertRole_POST_v1_roles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "UpsertRole",
        "x-source-file": "internal/apiserver/handler/rbac.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role_id": {
                    "type": "string"
                  },
                  "display_name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/roles/{id}": {
      "delete": {
        "tags": [
          "RBAC"
        ],
        "summary": "DeleteRole",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `DeleteRole`.",
        "operationId": "DeleteRole_DELETE_v1_roles_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "DeleteRole",
        "x-source-file": "internal/apiserver/handler/rbac.go"
      },
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "GetRole",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `GetRole`.",
        "operationId": "GetRole_GET_v1_roles_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "GetRole",
        "x-source-file": "internal/apiserver/handler/rbac.go"
      },
      "put": {
        "tags": [
          "RBAC"
        ],
        "summary": "UpsertRole",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `UpsertRole`.",
        "operationId": "UpsertRole_PUT_v1_roles_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "UpsertRole",
        "x-source-file": "internal/apiserver/handler/rbac.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role_id": {
                    "type": "string"
                  },
                  "display_name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/roles/{id}/permissions": {
      "post": {
        "tags": [
          "RBAC"
        ],
        "summary": "AssignRolePermissions",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `AssignRolePermissions`.",
        "operationId": "AssignRolePermissions_POST_v1_roles_id_permissions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "AssignRolePermissions",
        "x-source-file": "internal/apiserver/handler/rbac.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/users": {
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "ListUsers",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `ListUsers`.",
        "operationId": "ListUsers_GET_v1_users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "ListUsers",
        "x-source-file": "internal/apiserver/handler/rbac.go"
      },
      "post": {
        "tags": [
          "RBAC"
        ],
        "summary": "UpsertUser",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `UpsertUser`.",
        "operationId": "UpsertUser_POST_v1_users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "UpsertUser",
        "x-source-file": "internal/apiserver/handler/rbac.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "team_id": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/users/{id}/roles": {
      "post": {
        "tags": [
          "RBAC"
        ],
        "summary": "AssignUserRoles",
        "description": "Auto-generated from handler route registration. RBAC scope/action hint: `rbac.admin`. Handler: `AssignUserRoles`.",
        "operationId": "AssignUserRoles_POST_v1_users_id_roles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-rbac-scope": "rbac.admin",
        "x-handler": "AssignUserRoles",
        "x-source-file": "internal/apiserver/handler/rbac.go",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "reason": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  }
}